refactor: Organize DDL AST definitions into logical modules
Split the 560-line `crates/ast/src/ddl.rs` into focused modules by feature category:

- **table.rs**: Table operations (CREATE/DROP/ALTER TABLE) and constraints
- **transaction.rs**: Transaction control (BEGIN/COMMIT/ROLLBACK/SAVEPOINT)
- **schema.rs**: Schema, view, index, role, and catalog operations
- **cursor.rs**: Cursor operations (DECLARE/OPEN/FETCH/CLOSE)
- **advanced.rs**: Advanced SQL:1999 objects (SEQUENCE, TYPE, DOMAIN, COLLATION, etc.)
- **mod.rs**: Re-exports all types for backward compatibility

Benefits:
- Logical organization by DDL category
- Easier navigation (find CREATE TABLE in table.rs, not line 16 of 560)
- Module-level documentation for each DDL category
- No breaking changes to public API (re-exports maintain compatibility)
- Zero runtime impact (organization only)

All 716 parser tests pass, confirming the refactoring maintains existing functionality.

Closes #787
