+++ crates/ast/src/ddl/table.rs
@@ -1,0 +1,141 @@
+//! Table DDL operations
+//!
+//! This module contains AST nodes for table-related DDL operations:
+//! - CREATE TABLE
+//! - DROP TABLE
+//! - ALTER TABLE (add/drop column, add/drop constraint, etc.)
+
+use crate::Expression;
+use types::DataType;
+
+/// Referential action for foreign key constraints
+#[derive(Debug, Clone, PartialEq)]
+pub enum ReferentialAction {
+    NoAction,
+    Cascade,
+    SetNull,
+    SetDefault,
+}
+
+/// CREATE TABLE statement
+#[derive(Debug, Clone, PartialEq)]
+pub struct CreateTableStmt {
+    pub table_name: String,
+    pub columns: Vec<ColumnDef>,
+    pub table_constraints: Vec<TableConstraint>,
+}
+
+/// Column definition
+#[derive(Debug, Clone, PartialEq)]
+pub struct ColumnDef {
+    pub name: String,
+    pub data_type: DataType,
+    pub nullable: bool,
+    pub constraints: Vec<ColumnConstraint>,
+    pub default_value: Option<Box<Expression>>,
+}
+
+/// Column-level constraint
+#[derive(Debug, Clone, PartialEq)]
+pub struct ColumnConstraint {
+    pub name: Option<String>,
+    pub kind: ColumnConstraintKind,
+}
+
+/// Column constraint types
+#[derive(Debug, Clone, PartialEq)]
+pub enum ColumnConstraintKind {
+    NotNull,
+    PrimaryKey,
+    Unique,
+    Check(Box<Expression>),
+    References {
+        table: String,
+        column: String,
+        on_delete: Option<ReferentialAction>,
+        on_update: Option<ReferentialAction>,
+    },
+}
+
+/// Table-level constraint
+#[derive(Debug, Clone, PartialEq)]
+pub struct TableConstraint {
+    pub name: Option<String>,
+    pub kind: TableConstraintKind,
+}
+
+/// Table constraint types
+#[derive(Debug, Clone, PartialEq)]
+pub enum TableConstraintKind {
+    PrimaryKey {
+        columns: Vec<String>,
+    },
+    ForeignKey {
+        columns: Vec<String>,
+        references_table: String,
+        references_columns: Vec<String>,
+        on_delete: Option<ReferentialAction>,
+        on_update: Option<ReferentialAction>,
+    },
+    Unique {
+        columns: Vec<String>,
+    },
+    Check {
+        expr: Box<Expression>,
+    },
+}
+
+/// DROP TABLE statement
+#[derive(Debug, Clone, PartialEq)]
+pub struct DropTableStmt {
+    pub table_name: String,
+    pub if_exists: bool,
+}
+
+/// ALTER TABLE statement
+#[derive(Debug, Clone, PartialEq)]
+pub enum AlterTableStmt {
+    AddColumn(AddColumnStmt),
+    DropColumn(DropColumnStmt),
+    AlterColumn(AlterColumnStmt),
+    AddConstraint(AddConstraintStmt),
+    DropConstraint(DropConstraintStmt),
+}
+
+/// ADD COLUMN operation
+#[derive(Debug, Clone, PartialEq)]
+pub struct AddColumnStmt {
+    pub table_name: String,
+    pub column_def: ColumnDef,
+}
+
+/// DROP COLUMN operation
+#[derive(Debug, Clone, PartialEq)]
+pub struct DropColumnStmt {
+    pub table_name: String,
+    pub column_name: String,
+    pub if_exists: bool,
+}
+
+/// ALTER COLUMN operation
+#[derive(Debug, Clone, PartialEq)]
+pub enum AlterColumnStmt {
+    SetDefault { table_name: String, column_name: String, default: Expression },
+    DropDefault { table_name: String, column_name: String },
+    SetNotNull { table_name: String, column_name: String },
+    DropNotNull { table_name: String, column_name: String },
+}
+
+/// ADD CONSTRAINT operation
+#[derive(Debug, Clone, PartialEq)]
+pub struct AddConstraintStmt {
+    pub table_name: String,
+    pub constraint: TableConstraint,
+}
+
+/// DROP CONSTRAINT operation
+#[derive(Debug, Clone, PartialEq)]
+pub struct DropConstraintStmt {
+    pub table_name: String,
+    pub constraint_name: String,
+}
